account_api/pyproject.toml
[project]
name = "account-api"
version = "0.1.0"
description = "Bank account API"
authors = [
    {name = "Samuel-Work",email = "patriciosamuel097@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
